fix build errors
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@36857 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/openssl/include/openssl/aes.h
@@ -36,7 +36,7 @@ struct aes_key_st {
 # endif
     int rounds;
 #ifdef OCTEON_OPENSSL
-    uint64_t cvmkey[4];
+    unsigned long long cvmkey[4];
     int cvm_keylen;
 #endif
 };
